+++ app/src/main/java/com/example/testtask/domain/usecase/GetBriefCasesUseCase.kt
@@ -4,7 +4,7 @@ import com.example.testtask.domain.repository.BriefCaseRepository
 
 class GetBriefCasesUseCase (private val briefCaseRepository: BriefCaseRepository){
 
-    fun execute (id: Int){
+    fun execute (id: Long){
         briefCaseRepository.getBriefCase(id)
     }
 }
